requirements.txt
streamlit==1.30.0
python-dotenv==1.0.1
pyautogen==0.7.6
langgraph==0.3.0
google-genai==1.25.0
openai==1.66.2
requests==2.31.0
vertexai>=1.71.1
ag2==0.9.6